Fix insert fragment to use dialog moved into QtGui
